fix(planner): extract emoji selector into dedicated popover component
