+++ src/components/planner/sidebar/selectedOptionController/index.ts
@@ -1,5 +1,6 @@
 import CopyOption from './CopyOption'
+import EmojiSelector from './EmojiSelector'
 import PasteOption from './PasteOption'
 import RandomFill from './RandomFill'
 
-export { CopyOption, PasteOption, RandomFill }
+export { CopyOption, EmojiSelector, PasteOption, RandomFill }
